Experience and Dimensions
Created an Experience class, which should hopefully make controlling
experience for the multiple dimensions smooth. Also created a Dimensions
enum. The Experience class has a static factory type function in order
to instantiate a default Experience, which adds all dimensions from the
enum with an initial value of zero. Testing for this needs to be done.