Fixed Experience.DefaultExperience() and added Testing for it
Experience.DefaultExperience() now returns an experience using the
Dimensions enum and the descriptions of each value therein. In this way
the DefaultExperience will not need to be modified in order to be used
to represent the user's exp. The Descriptions were used because the
database is storing the Dimensions strings with spaces.