Bug fix for 3 of the Dimensions not gaining EXP
The strings used to refer to the dimensions were different from the ones
used within the database, and so they weren't being updated properly.
The descriptions of the Dimension enum were fixed to match the strings
in the database. Reference to exact string names were replaced with
references to the description value in order to reduce the need to make
string changes in multiple areas.